# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UKM Politeknik",
        "description": "API UKM Politeknik Negeri Lampung. Response dapat berubah sewaktu waktu menyesuaikan perbaikan",
        "contact": {
            "email": "leopradana013@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Production"
        }
    ],
    "paths": {
        "/agenda/highlight": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Menampilkan 1 agenda terbaru",
                "operationId": "c1f322a323b787922c90b5b1e85ef726",
                "parameters": [
                    {
                        "parameter": "id_ukm",
                        "name": "id_ukm",
                        "in": "query",
                        "description": "ID UKM",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "example nama"
                                                    },
                                                    "jenis": {
                                                        "type": "enum",
                                                        "example": "example jenis"
                                                    },
                                                    "singkatan_ukm": {
                                                        "type": "string",
                                                        "example": "example singkatan_ukm"
                                                    },
                                                    "foto_ukm": {
                                                        "type": "text",
                                                        "example": "example foto_ukm"
                                                    },
                                                    "keterangan": {
                                                        "type": "string",
                                                        "example": "example keterangan"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/agenda/detail/{agenda_id}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Menampilkan detail agenda",
                "operationId": "0bdab274cb411aa2449f0b281afdfbad",
                "parameters": [
                    {
                        "parameter": "agenda_id",
                        "name": "agenda_id",
                        "in": "path",
                        "description": "ID Agenda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "id_ukm",
                        "name": "id_ukm",
                        "in": "query",
                        "description": "ID UKM",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "example nama"
                                                    },
                                                    "jenis": {
                                                        "type": "enum",
                                                        "example": "example jenis"
                                                    },
                                                    "singkatan_ukm": {
                                                        "type": "string",
                                                        "example": "example singkatan_ukm"
                                                    },
                                                    "foto_ukm": {
                                                        "type": "text",
                                                        "example": "example foto_ukm"
                                                    },
                                                    "keterangan": {
                                                        "type": "string",
                                                        "example": "example keterangan"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}